Status, Priority and Service Category Filters
On the page of Services -> Service Board and Site Admin -> Admin Service
Board, for the Status filter, currently, it is a text box, we are
thinking it might be more user friendly to change it to a dropdown list,
so users can select from the list.
The same change of from a text box to a dropdown list for Priority
filter and Service Category filter on the same page, and status filter
on Services -> Bid Requests Board page as well.