Priority and Category Editing Capabilities and Bid Request
corp admin needs to be able to edit priority level & service category.
Can we make a dropdown menu?
Manuel:Estimate is 3.5 hours

corp admin needs to be able to enter a new bid request for customer
(same as a service request).  Need to be able to create bid requests for
all locations.
Manuel:Estimate is 4 hours